Moved tests for transferring playback to SingleDeviceOnlyTransferPlaybackCommandHandlerDelegateTest instead testing it directly in DefaultStorageDeviceOperationsTest
